+++ ingenialink/ethernet/register.py
@@ -50,7 +50,6 @@ class EthernetRegister(Register):
         internal_use=0,
         address_type=None,
     ):
-
         super().__init__(
             dtype,
             access,