+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Changed
+- Update the load_FWs script to only use ingenialink.
 - Remove binding to the [ingenialink](https://github.com/ingeniamc/ingenialink) C library.
 
 ## [6.5.1] - 2023-01-17