+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Update the load_FWs script to only use ingenialink.
+- Improve the load_FWs script when using canopen protocol.
 - Remove binding to the [ingenialink](https://github.com/ingeniamc/ingenialink) C library.
 
 ## [6.5.1] - 2023-01-17