+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Improve the load_FWs script when using canopen protocol.
 - Remove binding to the [ingenialink](https://github.com/ingeniamc/ingenialink) C library.
 
+###Fixed
+- NACK error code formatting.
+
 ## [6.5.1] - 2023-01-17
 ### Fixed
 - Truncate received data in ethernet by the expected data length.